refactor: PointArticle 컴포넌트 분리 및 매직 넘버 상수화
