+++ components/PointArticle.module.scss
@@ -1,0 +1,30 @@
+.point-article {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+}
+
+.point-number {
+  font-weight: 700;
+  font-size: 18px;
+  color: #4a494f;
+}
+
+.explanation-container {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.explanation {
+  font-weight: 700;
+  font-size: 24px;
+  line-height: 36px;
+  color: #1a1a1a;
+}
+
+.sub-explanation {
+  font-size: 16px;
+  line-height: 26px;
+  color: #555555;
+}
